HBI: Refactor WelcomeModal, Fix Links in Modal, Fix Overlay-Layer visibility (#954)
* WelcomeModal: Extract styles in separate file

* WelcomeModal: Split Components, Fix Column2 Center/Zoom

- Center/Zoom for Column2 does not reload the page anymore, so the closed-State of the modal is preserved.
- Column1 still has a reload, though. But this is less annoying.
- Columns are split into Components to keep the Modal Component small.

* Map: Bugfix: Overlay layers where not filtered properly

Fixes #934
