+++ src/pages/Reports/pages/OverviewMap/components/WelcomeModal/GridChilds/index.ts
@@ -1,0 +1,3 @@
+export { Column1 } from './Column1';
+export { Column2 } from './Column2';
+export { Column3 } from './Column3';
